Make a component for cart lines to easily update
